fix: Reviewers comments
Move overwrite tests to its own files
